+++ www/src/components/Plural.js
@@ -7,6 +7,7 @@ import BreadcrumbProvider from './Breadcrumbs'
 import Chart from './repos/Chart'
 import EditUser from './users/EditUser'
 import Marketplace from './marketplace/Marketplace'
+import OAuthCallback from './shell/OAuthCallback'
 import Repository from './repository/Repository'
 import RepositoryArtifacts from './repository/RepositoryArtifacts'
 import RepositoryDeployments from './repository/RepositoryDeployments'
@@ -19,7 +20,7 @@ import RepositoryPackagesTerraform from './repository/RepositoryPackagesTerrafor
 import RepositoryTests from './repository/RepositoryTests'
 import Terraform from './repos/Terraform'
 import { Audits } from './accounts/Audits'
-import { CloudShell, OAuthCallback } from './shell/CloudShell'
+import CloudShell from './shell/CloudShell'
 import { CurrentUserContext, PluralConfigurationContext, PluralProvider } from './login/CurrentUser'
 import { DeviceLoginNotif } from './users/DeviceLoginNotif'
 import { Docker, DockerRepository } from './repos/Docker'
