+++ www/src/components/shell/onboarding/OnboardingHeader.js
@@ -1,0 +1,18 @@
+import { Box } from 'grommet'
+import { Text } from 'honorable'
+
+export default function Header({ text }) {
+  return (
+    <Box
+      fill="horizontal"
+      align="center"
+      pad="small"
+    >
+      <Text
+        size="small"
+        weight={500}
+      >{text}
+      </Text>
+    </Box>
+  )
+}
