+++ www/src/components/shell/onboarding/OnboardingNavSection.js
@@ -1,0 +1,11 @@
+import { Flex } from 'honorable'
+
+export default function OnboardingNavSection(props) {
+  return (
+    <Flex
+      marginTop="xxlarge"
+      justify="space-between"
+      {...props}
+    />
+  )
+}
