+++ www/src/components/shell/onboarding/synopsis/Synopsis.js
@@ -1,0 +1,94 @@
+import { Box } from 'grommet'
+import { Text } from 'honorable'
+
+import { synopsis } from '../cloud/provider'
+
+function SynopsisTable({ items, width }) {
+  return (
+    <Box
+      gap="xsmall"
+      border={{ side: 'between' }}
+    >
+      {items.map(({ name, value }) => (
+        <Box
+          direction="row"
+          key={name}
+          gap="small"
+          align="center"
+        >
+          <Box
+            flex={false}
+            width={width || '120px'}
+          >
+            <Text
+              size="small"
+              weight={500}
+            >
+              {name}
+            </Text>
+          </Box>
+          <Box fill="horizontal">{value}</Box>
+        </Box>
+      ))}
+    </Box>
+  )
+}
+
+function SynopsisSection({ header, children }) {
+  return (
+    <Box
+      pad="small"
+      round="xsmall"
+      background="card"
+      gap="xsmall"
+    >
+      <Box
+        direction="row"
+        justify="center"
+      >
+        <Text
+          size="small"
+          weight={500}
+        >{header}
+        </Text>
+      </Box>
+      {children}
+    </Box>
+  )
+}
+
+function Synopsis({ scm, credentials, workspace, provider, demo }) {
+  return (
+    <Box gap="small">
+      <Text size="small">You've completed the configuration for your shell, but let's verify everything looks good just to be safe</Text>
+      <Box
+        gap="medium"
+        direction="row"
+      >
+        <SynopsisSection header="Git Setup">
+          <SynopsisTable
+            width="80px"
+            items={[{ name: 'Repository', value: scm.org ? `${scm.org}/${scm.name}` : scm.name }]}
+          />
+        </SynopsisSection>
+        {!demo && (
+          <SynopsisSection header="Cloud Credentials">
+            <SynopsisTable items={synopsis({ provider, credentials, workspace })} />
+          </SynopsisSection>
+        )}
+        <SynopsisSection header="Workspace">
+          <SynopsisTable
+            width="100px"
+            items={[
+              { name: 'Cluster', value: workspace.cluster },
+              { name: 'Bucket Prefix', value: workspace.bucketPrefix },
+              { name: 'Subdomain', value: workspace.subdomain },
+            ]}
+          />
+        </SynopsisSection>
+      </Box>
+    </Box>
+  )
+}
+
+export default Synopsis
